fix typings
